Switch to the more modern `thiserror` crate
This commit removes the dependency on `failure` and `failure_derive` in
favor of the more modern `thiserror` crate. This is intended to
hopefully bring build-time wins to wasmtime because `failure_derive`
depends on `synstructure` which enables lots of features of `syn`. This
means that if you alternate between `cargo build` and `cargo test`
features are changing and lots of rebuilds happen when this crate is
conditionally depended on in tests (as it often is).